pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "metamorfose-whatsapp-chatbot"
version = "0.1.0"
description = "FastAPI backend scaffold for Metamorfose WhatsApp chatbot."
authors = [
  { name = "Felipe Kelemen Bernardino", email = "felipe.kelemen@usp.br" },
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "sqlalchemy",
  "fastapi",
  "uvicorn[standard]",
  "pydantic-settings",
  "python-dotenv",
  "sqlalchemy",
  "httpx",
  "redis>=7.3.0",
  "python-telegram-bot>=22.7",
  "dotenv>=0.9.9",
  "types-redis>=4.6.0.20241004",
  "toml>=0.10.2",
  "types-toml>=0.10.8.20260408",
  "psycopg[binary]"
]

[project.scripts]
app = "app.__main__:main"

[tool.setuptools.packages.find]
include = ["app"]

[dependency-groups]
dev = ["pytest", "pytest-asyncio", "ruff", "mypy", "pre-commit", "autopep8>=2.3.2", "build"]

[tool.ruff]
line-length = 100
target-version = "py311"
exclude = ["__pycache__", ".venv", "build", "dist", ".eggs"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "Q", "S", "B", "C4", "TID"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101", "S106", "S105"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["."]
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true

[[tool.mypy.overrides]]
module = "tests.*"
disable_error_code = ["no-untyped-def", "var-annotated", "attr-defined"]
